recent_summoners function to aggregate calls
Combine all the calls built up to this point, chaining them to assemble
a list of names of all summoners the provided summoner has played with
in the last five matches.

When collecting the match information, parallelized into five separate
tasks to speed up the response, since they are each independent.