Begin migration to supervised app and Tesla
Initial pass up to this point was exploring using straight `:httpc`,
since it is built into BEAM. I had some thoughts about how to test this,
but didn't like the direction that was heading, so also switching to
Tesla here shortly to make life a bit easier.

Also setting the path that will break in a different way when the Riot
Games API key is not set.